refactor: Move LegMatcher and ItineraryMatcher to itinerarymatching pkg.
